requirements.txt
opencv-contrib-python>=4.5
numpy>=1.22
scipy>=1.7
pillow>=9.0
pandas>=1.4
